backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
opencv-python
Pillow
python-multipart
